Refactor ModalTask and SaveEditedTask components to add a save message
Previously, the ModalTask component had a SaveEditedTask component that only had a button to save the edited task. Now, the SaveEditedTask component also has a save message to inform the user of the success or failure of the task update.

The ModalTask component now includes the save message after the save button, which appears when the task has been successfully saved. The save message disappears after 500 milliseconds.

Additionally, the SaveEditedTask component now takes a setSaveMessage prop, which is used to set the save message upon success or failure of the task update.