Add QuickEditPriority component for updating task priority
The QuickEditPriority component allows users to update the priority of a task. It includes a form with a dropdown menu for selecting the new priority. After submitting the form, the priority of the task is updated in the store and the changes are reflected in the UI. If the update is successful, a success toast is displayed. If there is an error, an error toast is displayed.