Refactor QuickEditStatus component to use setExclusiveEditingField
This commit refactors the QuickEditStatus component in order to utilize the setExclusiveEditingField action instead of the previous setEditingField action.
By using setExclusiveEditingField, we can ensure that only one field is being edited at a time, preventing conflicts in the editing state.
This change improves the user experience by providing a more intuitive and seamless editing workflow.
No known issues or side effects are anticipated as a result of this change.